Move setSecondSecret function to basic.js file
