+++ lib/transactions/basic.js
@@ -1,3 +1,4 @@
+var crypto = require("./crypto.js")
 var transaction = require('./transaction.js')
 
 function calculateFee(username){
@@ -26,6 +27,25 @@ function setName(username, secret, secondSecret) {
 	})
 }
 
+function newSignature(secondSecret) {
+	var keys = crypto.getKeys(secondSecret);
+	var signature = {
+		publicKey: keys.publicKey
+	};
+	return signature;
+}
+
+function setSecondSecret(secret, secondSecret) {
+	var secondSignature = newSignature(secondSecret);
+	return transaction.createTransactionEx({
+		type: 3,
+		fee: 5 * 1e8,
+		args: [secondSignature.publicKey],
+		secret: secret
+	})
+}
+
 module.exports = {
-	setName: setName
+	setName: setName,
+	setSecondSecret: setSecondSecret
 }
